feat: addd api routes for products in ApiController
