+++ .gitignore
@@ -15,3 +15,5 @@
 
 ## Jetbrains IDEs
 .idea
+
+.vscode
