+++ README.md
@@ -5,10 +5,29 @@ CyberMatos is a toy e-commerce site built with Symfony and deployed with Ansible
 # Setup
 
 1. Import the cybermatos.sql file into a mariaDB database.
-2. Check that the DATABASE_URL environment variable in the .env is: "mysql://user:mdp@127.0.0.1:3307/cybermatos?serverVersion=10.6.5-MariaDB"
+2. Create a file `.en.local` with environment variable : `DATABASE_URL="mysql://user:password@127.0.0.1:3307/cybermatos?serverVersion=10.6.5-MariaDB"`
 3. Install dependencies with `composer install`
 4. Start the server with `symfony server:start`
-5. Go to homepage -> https://localhost:34795
+5. Go to homepage `https://localhost:+port` with the link specified in the logs.
+
+# Api routes
+
+List of API routes with JSON content type
+
+Retrieve a product or the list of products
+
+✔️ **/api/products**   
+✔️ **/api/products/{productId}**
+
+```
+{
+" id ": 1 ,
+" name ": " Item 3000" ,
+" description ": " Best item in the shop !" ,
+" photo ": " https :// path / to / image . png " ,
+" price ": 13.37
+}
+```
 
 ### Entity Relationship Diagrams
