+++ app/src/Controller/ProductController.php
@@ -7,31 +7,12 @@ use App\Form\ProductType;
 use App\Repository\ProductRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 #[Route('/products')]
 class ProductController extends AbstractController
 {
-    #[Route('/', name: 'app_product_index', methods: ['GET'])]
-    public function index(ProductRepository $productRepository): Response
-    {
-        $products = $productRepository->findAll();
-        $jsonData = [];
-        foreach ($products as $product) {
-            $jsonData[] = [
-                'id' => $product->getId(),
-                'name' => $product->getName(),
-                'description' => $product->getDescription(),
-                'photo' => $product->getImage(),
-                'price' => $product->getPrice()
-            ];
-        }
-
-        return $this->json($jsonData);
-    }
-
     #[Route('/list', name: 'app_product_list', methods: ['GET'])]
     public function list(ProductRepository $productRepository): Response
     {
@@ -58,21 +39,6 @@ class ProductController extends AbstractController
             'product' => $product,
             'form' => $form,
         ]);
-    }
-
-
-    #[Route('/{id}', name: 'app_product_json', methods: ['GET'])]
-    public function productId(Product $product): Response
-    {
-        $jsonData = [
-            'id' => $product->getId(),
-            'name' => $product->getName(),
-            'description' => $product->getDescription(),
-            'photo' => $product->getImage(),
-            'price' => $product->getPrice()
-        ];
-
-        return $this->json($jsonData);
     }
 
     #[Route('/id/{id}', name: 'app_product_show', methods: ['GET'])]
