+++ app/templates/product/list.html.twig
@@ -42,5 +42,4 @@
 	</table>
 
 	<a href="{{ path('app_product_new') }}">Créer un produit</a><br>
-	<a href="{{ path('app_product_index') }}">{JSON}</a>
 {% endblock %}
